// sales-management-app/src/types/company.ts
export interface CompanyInfo {
  company_name: string;
  company_postalCode: string;
  company_address: string;
  company_phone: string;
  company_fax: string;
  company_mail: string;
  company_contactPerson: string;
  company_bankName: string;
  company_bankBranch: string;
  company_bankType: string;
  company_bankNumber: string;
  company_bankHolder: string;
  company_invoiveNumber: string;
}